% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqrmv.R
\name{dqrmvnorm}
\alias{dqrmvnorm}
\alias{dqrmvt}
\title{Multivariate Distributions}
\usage{
dqrmvnorm(n, ...)

dqrmvt(n, ...)
}
\arguments{
\item{n}{number of  observations}

\item{...}{forwarded to \code{\link[mvtnorm]{rmvnorm}} or \code{\link[mvtnorm]{rmvt}}}
}
\value{
numeric matrix of multivariate normal or t-distributed variables
}
\description{
Multivariate Distributions
}
\examples{
sigma <- matrix(c(4,2,2,3), ncol=2)
x <- dqrmvnorm(n=500, mean=c(1,2), sigma=sigma)
colMeans(x)
var(x)
plot(x)
## X ~ t_3(0, diag(2))
x <- dqrmvt(100, sigma = diag(2), df = 3) # t_3(0, diag(2)) sample
plot(x)
}
\seealso{
\code{\link[mvtnorm]{rmvnorm}} and \code{\link[mvtnorm]{rmvt}}
}
