% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqrng-package.R
\docType{package}
\name{dqrng}
\alias{dqrng}
\alias{dqrng-package}
\title{Fast Pseudo Random Number Generators for R}
\description{
This package provides several fast random number generators as C++
header only libraries. In addition fast functions for generating random
numbers according to a uniform, normal and exponential distribution are
provided. These functions are exported to R and as a C++ interface.
}
