% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRawSeeds.R
\name{generateRawSeeds}
\alias{generateRawSeeds}
\title{Generate seed as a raw vector}
\usage{
generateRawSeeds(n, bits)
}
\arguments{
\item{n}{Integer scalar, number of seeds to generate.}

\item{bits}{Integer scalar, number of bits to store per seed.}
}
\value{
A list of length \code{n}, where each element is a raw vector that
contains \code{bits} bits of randomness.
}
\description{
Generate seed as a raw vector
}
\details{
Each seed is encoded as a raw vector with the most significant bits
at the start of the vector. Each raw vector is converted into an
unsigned integer by the following procedure:
\enumerate{
\item Start with a sum of zero.
\item Add the first value of the vector.
\item Left-shift the sum by 8.
\item Add the next value of the vector, and repeat.
}

The aim is to facilitate R-level generation of seeds with sufficient 
randomness to cover the entire state space of pseudo-random number 
generators that require more than the ~32 bits available in a typical
\code{int}. It also preserves the integer nature of the seed, thus
avoiding problems with casting double-precision numbers to integers.
}
\examples{
generateRawSeeds(10, 64)

generateRawSeeds(5, 128)

generateRawSeeds(10, 50) # does not need to be divisible by 8.

}
\author{
Aaron Lun
}
