% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqrmv.R
\name{dqrmvnorm}
\alias{dqrmvnorm}
\title{Multivariate Distributions}
\usage{
dqrmvnorm(n, ...)
}
\arguments{
\item{n}{number of  observations}

\item{...}{forwarded to \code{\link[mvtnorm]{rmvnorm}}}
}
\value{
numeric matrix of multivariate normal distributed variables
}
\description{
Multivariate Distributions
}
\examples{
\dontshow{if (requireNamespace("mvtnorm", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sigma <- matrix(c(4,2,2,3), ncol=2)
x <- dqrmvnorm(n=500, mean=c(1,2), sigma=sigma)
colMeans(x)
var(x)
plot(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[mvtnorm]{rmvnorm}}
}
