% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dqrng-package.R
\name{dqrng functions}
\alias{dqrng functions}
\alias{dqset.seed}
\alias{dqRNGkind}
\alias{dqrunif}
\alias{dqrnorm}
\alias{dqrexp}
\alias{dqrng functions}
\title{dqrng functions}
\usage{
dqset.seed(seed)

dqRNGkind(kind, normal_kind = "ignored")

dqrunif(n, min = 0, max = 1)

dqrnorm(n, mean = 0, sd = 1)

dqrexp(n, rate = 1)
}
\arguments{
\item{seed}{integer seed for the random number generator}

\item{kind}{string specifying the RNG (see details)}

\item{normal_kind}{ignored; included for compatibility with \code{\link{RNGkind}}}

\item{n}{number of  observations}

\item{min}{lower limit of the uniform distribution}

\item{max}{upper limit of the uniform distribution}

\item{mean}{mean value of the normal distribution}

\item{sd}{standard deviation of the normal distribution}

\item{rate}{rate of the exponential distribution}
}
\value{
\code{dqrunif}, \code{dqrnorm}, and \code{dqrexp} return a numeric vector of length \code{n}.
}
\description{
The \code{dqrng} package provides function for random number generation.
These functions are modeled after the \code{base} functions \code{\link{set.seed}},
\code{\link{RNGkind}}, \code{\link{runif}}, \code{\link{rnorm}}, and \code{\link{rexp}}.
}
\details{
Supported RNG kinds:
\describe{
  \item{Mersenne-Twister}{The 64bit variant of Mersenne-Twister. Also the default.}
  \item{pcg64}{\url{http://www.pcg-random.org}}
  \item{Xorshift128+}{\url{http://xoroshiro.di.unimi.it/}}
  \item{Xorshift1024*}{\url{http://xoroshiro.di.unimi.it/}}
  \item{Xoroshiro128+}{\url{http://xoroshiro.di.unimi.it/}}
}

\code{rnorm} and \code{rexp} use the Ziggurat method as provided by \code{boost.random}.
}
\seealso{
\code{\link{set.seed}}, \code{\link{RNGkind}}, \code{\link{runif}},
         \code{\link{rnorm}}, and \code{\link{rexp}}
}
